Fix void-pointer-to-enum-cast warnings
ExmString.c:
- Replace (XmTextType)NULL with proper enum value XmNO_TEXT
- Fixes casting NULL to enum type in XmStringUnparse call

piano.c:
- Fix (NoteType)XtPointer casts by using intermediate long cast
- (NoteType)noteNumber -> (NoteType)(long)noteNumber
- (NoteType)clientData -> (NoteType)(long)clientData
- Prevents undefined behavior when casting pointers to smaller enum types

+++ demos/programs/piano/piano.c
@@ -915,7 +915,7 @@ void SoundCB (Widget w, XtPointer noteNumber, XtPointer callData)
     {
       note.display      = XtDisplay(w);
       note.noteType     = EIGHTH;
-      note.noteNumber   = (NoteType)noteNumber;
+      note.noteNumber   = (NoteType)(long)noteNumber;
       note.noteIndex    = 0;
       note.ledgerLine   = 0;
       note.noteDuration = appData->baseDuration;
@@ -1144,7 +1144,7 @@ void SetActiveNote (Widget w, NoteType noteType)
  *--------------------------------------------------------------------*/
 void SetNoteCB (Widget w, XtPointer clientData, XtPointer callData)
 {
-   NoteType noteType = (NoteType)clientData;
+   NoteType noteType = (NoteType)(long)clientData;
 
    SetActiveNote(w, noteType);
 }
